+++ packages/bug-repro-13/src/bar.ts
@@ -1,0 +1,3 @@
+export type Bar = {
+  x: number;
+};